requirements.txt
# Bybit Trading Bot Dependencies
# Python 3.9+

# Bybit API
pybit>=5.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Technical Indicators
pandas-ta>=0.3.14b

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.8.0
aiosqlite>=0.17.0

# Notifications
python-telegram-bot>=20.0

# Utilities
cryptography>=41.0.0
